Support progressive completion of password items
We complete only the current level matched (including trailing slash) when the filter string is a prefix of the matched item.

For example if the item is `foo/bar/baz` and we type `f⇥` then it gets completed to `foo/` and pressing tab again will make it `foo/bar/` etc.